FastAPI/cloud/requirements.txt
# FastAPI Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.1

# Database Dependencies
SQLAlchemy==2.0.45
alembic==1.18.4
sqlmodel==0.0.14

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.10.1
pydantic-settings==2.10.1
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9
limits==5.8.0

# HTTP Client
httpx==0.27.2
requests==2.31.0

# Email (for background tasks)
# Note: Add your email service dependencies here if needed
# For SMTP: aiosmtplib
# For SendGrid: sendgrid
# For AWS SES: boto3

# Development & Testing Dependencies
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==4.1.0

# Optional: Production Database Drivers
# Uncomment if switching from SQLite
# psycopg2-binary==2.9.7  # PostgreSQL
# pymysql==1.1.0         # MySQL

# Optional: Caching
# redis==4.5.4
# aioredis==2.0.1

# Optional: Advanced Logging
# loguru==0.7.3

# Optional: Background Tasks
# celery==5.3.0
# redis==4.5.4

# Optional: API Documentation (already included with FastAPI)
# No additional dependencies needed for Swagger/OpenAPI
